refector: 패키지 rename ordertable -> table
