+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.tablegroup.dto;
 
-import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.tablegroup.domain.TableGroup;
 
 import java.time.LocalDateTime;
